==> 01_download.ipynb <==
# Download

Download the source data files from the [U.S. Department of Labor's Office of Foreign Labor Certification](https://www.foreignlaborcert.doleta.gov/performancedata.cfm).

!curl -o input/2017.xlsx https://www.foreignlaborcert.doleta.gov/pdf/PerformanceData/2017/H-2A_Disclosure_Data_FY17.xlsx
!curl -o input/2016.xlsx https://www.foreignlaborcert.doleta.gov/docs/Performance_Data/Disclosure/FY15-FY16/H-2A_Disclosure_Data_FY16.xlsx
!curl -o input/2015.xlsx https://www.foreignlaborcert.doleta.gov/docs/py2015q4/H-2A_Disclosure_Data_FY15_Q4.xlsx
!curl -o input/2014.xlsx https://www.foreignlaborcert.doleta.gov/docs/py2014q4/H-2A_FY14_Q4.xlsx
!curl -o input/2013.xlsx https://www.foreignlaborcert.doleta.gov/docs/h_2a/H2A_FY2013.xls
!curl -o input/2012.xlsx https://www.foreignlaborcert.doleta.gov/docs/h_2a/H-2A_FY2012.xlsx
!curl -o input/2011.xlsx https://www.foreignlaborcert.doleta.gov/docs/h_2a/H-2A_FY2011.xlsx
!curl -o input/2010.xlsx https://www.foreignlaborcert.doleta.gov/docs/h_2a/H-2A_FY2010.xlsx
!curl -o input/2009.xlsx https://www.foreignlaborcert.doleta.gov/docs/h_2a/H2A_FY2009.xlsx
!curl -o input/2008.xlsx https://www.foreignlaborcert.doleta.gov/docs/h_2a/H2A_FY2008.xlsx
